Split client side components into their own file
